requirements.txt
streamlit
openai
python-dotenv
openai-whisper @ git+https://github.com/openai/whisper.git
